+++ webpack.test.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-    mode: 'production',
+    mode: 'none',
     output: {
         filename: 'index.js',
         libraryTarget: 'umd',
@@ -16,7 +16,7 @@ module.exports = {
         rules: [
             {
                 test: /\.js?$/,
-                exclude: /(node_modules|bower_components|vender_modules)/,
+                exclude: /(node_modules)/,
                 loader: 'babel-loader',
             },
         ],